% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_freezer.R
\name{classify_freezer}
\alias{classify_freezer}
\title{Classification of phasic and sustained freezers}
\usage{
classify_freezer(data_MR1, data_MR2, sex, MR, shifter)
}
\arguments{
\item{data_MR1}{dataframe containing freezing values MR1 of bins 13 - 24 as a character}

\item{data_MR2}{dataframe containing freezing values MR2 of bins 13 - 24 as a character}

\item{sex}{"female" or "male" specifying the sex}

\item{MR}{1 or 2 specifying the Memory Retrieval Session 1 or 2}

\item{shifter}{boolean argument specifying if shifters should be included in prediction}
}
\description{
Classification of phasic and sustained freezers
}
