% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freezingDataset.R
\docType{data}
\name{freezingDataset}
\alias{freezingDataset}
\title{World Health Organization TB data}
\format{
\subsection{\code{freezingDataset}}{

A data frame with 40 rows and 36 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
<>
}
\usage{
freezingDataset
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
