% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrain_model.R
\name{retrain_model}
\alias{retrain_model}
\title{Retraining utility of the PhenoFreeze Package to add more data to the training}
\usage{
retrain_model(new_data, sex, MR, training_iterations = 1000)
}
\arguments{
\item{new_data}{dataframe containing new freezing values MR1 of bins 13 - 24 as a character}

\item{sex}{"female" or "male" specifying the sex of the new data}

\item{MR}{1 or 2 specifying the Memory Retrieval Session 1 or 2 of the new data}

\item{training_iterations}{description}
}
\description{
Retraining utility of the PhenoFreeze Package to add more data to the training
}
