% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.phenotype.R
\name{predict.phenotype}
\alias{predict.phenotype}
\title{Classification of phasic and sustained freezers}
\usage{
\method{predict}{phenotype}(data, ..., sex, MR)
}
\arguments{
\item{data}{dataframe containing freezing values of bins 13 - 24}

\item{...}{further parameters}

\item{sex}{"female" or "male" specifying the sex}

\item{MR}{1 or 2 specifying the Memory Retrieval Session 1 or 2}
}
\description{
Classification of phasic and sustained freezers
}
